New ZED-based AHRS driver, integrated into NavPosMgr
The new driver is for a system that uses the ZEDM or ZED2 built-in IMU.

This is the second AHRS driver we support (after the LORD). The exercise of integrating it exposed some abstraction issues with the NavPosMgr and LORAHRSDriver, so those were resolved as part of this effort.

Still todo is applying the config-file RPY offset. This should be quite easy using the ROS tf2 quaternion transforms (which includes functions to move between RPY and quaternions). Still TBD is whether to do that in the driver or NavPosMgr... Lord version does it in the driver because the device, itself, supports this notion.